Add a first (not quite functional) pass at providing display adapters
The idea is that a display adapter is an object responsible for providing enough field config for the renderer to display it nicely (and in different ways, for different display adapters). For example, you could have a “select” display adapter for a “string” field so it showed up as a select box instead of a single-line text input.

The problem with this initial implementation is that the container of display adapters is only available at the top-level of the form. It isn’t passed around into the all the definition contexts (i.e. anything nested into blocks). To do this nicely, I think we actually need to take a different approach to building up the form definition altogether, which I’ll do in the next commits.