Stop coercing form post data in the output compiler
Now that we require a schema, we can depend on that for form post data type coercions. Avoiding the coercions here means we also avoid errors due to the naive approach we’d taken so far. This was actually manifested in the formalist-demo app as errors when trying to post the demo form with empty values for the date and date_time fields. Now we just pass every form post value through to the schema as the posted value, or a blank string, and let it coerce things more appropriately.