Move embedded form collection behavior into a mixin
Make the EmbeddedFormCollection include this mixin, which allows the behavior to be offered at either a class or object level, just like how dry-container itself.

Make the Registration a proper class instead of a struct subclass, and along with the mixin itself, add proper `#to_h` output to enable the collection to be consumed by the rich text area field.