Remove custom predicates container and embedded_forms_valid? predicate
Now that we’re not using the transformer in front of any kind of validation schema, we need this embedded_forms_valid? predicate to actually compile each of the rich text strings, and this turns out to be pretty straightforward to implement directly in your app:

```ruby
require "formalist/rich_text/embedded_form_compiler"
require "formalist/rich_text/validity_check"
require “my_app/validation/form"
require "admin/container"

module Admin
  module Validation
    class Form < MyApp::Validation::Form
      option :embedded_forms, Container["formalist.embedded_forms.collection"]

      def embedded_forms_valid?(string)
        compiler = Formalist::RichText::EmbeddedFormCompiler.new(embedded_forms)

        Formalist::RichText::ValidityCheck.new.(compiler.(string))
      end
    end

    def self.Form(&block)
      Dry::Validation.Schema(Form, &block)
    end
  end
end
```

So until we establish a really nice way to help with this, it’s better just to leave this to the app.