Add input processor to embedded form registrations
This allows data for the form to be processed before it is used to build a form AST for incorporation into the rich text AST before displaying it in the rich text editor.

Applications can use this feature to provide extra, necessary, application-level data associated to the form data. For example, in the case of an "image" embedded form, the application would need to provide a "thumbnail_url" for the image form to display the thumbnail for the image.