Use ivars instead of accessors when creating filled forms/elements
Some form or element subclasses may override attribute readers and expand upon the data kept in the ivar. In these cases, we want to preserve the ivar’s original data when making a new copy of a form or element object, otherwise we may encounter unexpected behaviour and crashes if the ivar is re-created based on an altered state returned from the reader.