Remove dry-types dependency from core gem
We'll be moving to a more dynamic form definition arrangement in future, where the gem lets you compose *any* kind of form with *any* kind of elements/attributes, and it'll be up to the renderers to make sure they're compatible.

Removing dry-types at this point also helps with people updating their apps between major dry-types versions.